[level 0] Title: 두 수의 나눗셈, Time: 0.01 ms, Memory: 4.21 MB -BaekjoonHub

+++ C/프로그래머스/0/120806. 두 수의 나눗셈/두 수의 나눗셈.c
@@ -4,10 +4,10 @@
 
 int solution(int num1,int num2)
 {
-    double a = num1;
-    double b = num2; 
+    float a = num1;
+    float b = num2; 
     
-	double result = a / b;
+	float result = a / b;
 
 	return (int)(result * 1000);
 }
